Forbedring i utlisting og visning av søknadsdetaljer
