+++ src/components/_common/form-details/FormDetails.module.scss
@@ -1,0 +1,8 @@
+.formDetails {
+    margin-bottom: 1.75rem;
+}
+
+.buttonContainer {
+    display: flex;
+    flex-direction: row;
+}
