+++ src/components/_common/form-details/FormDetailsItem.module.scss
@@ -5,9 +5,11 @@
     width: max-content;
 
     &:not(:last-child) {
-        margin-bottom: 1.75rem;
+        margin-right: 1rem;
     }
+
     @media #{common.$mq-screen-mobile} {
         width: 100%;
+        margin-right: 0;
     }
 }
