+++ src/components/parts/form-details/FormDetailsPart.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormDetailsProps } from 'types/component-props/parts/form-details';
 import { EditorHelp } from 'components/_editor-only/editor-help/EditorHelp';
-import { FormsList } from 'components/_common/form-details/FormsList';
+import { FormDetails } from 'components/_common/form-details/FormDetails';
 
 export const FormDetailsPart = ({ config }: FormDetailsProps) => {
     const { targetFormDetails } = config;
@@ -15,5 +15,5 @@ export const FormDetailsPart = ({ config }: FormDetailsProps) => {
         );
     }
 
-    return <FormsList formDetails={formDetails} />;
+    return <FormDetails formDetails={formDetails} />;
 };
